[FSDP()][9/N] Refactor ctor (continued)
[ghstack-poisoned]
